+++ mls/nostr-mls-memory-storage/src/lib.rs
@@ -250,6 +250,7 @@ mod tests {
         let nostr_group_id = create_test_nostr_group_id();
         let image_url = Some("http://blossom_server:4531/fake_img.png".to_owned());
         let image_key = Some(generate_encryption_key());
+        let image_nonce = Some(vec![16u8; 12]);
         let group = Group {
             mls_group_id: mls_group_id.clone(),
             nostr_group_id,
@@ -262,6 +263,7 @@ mod tests {
             state: GroupState::Active,
             image_url,
             image_key,
+            image_nonce,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let found_group = nostr_storage
@@ -290,6 +292,7 @@ mod tests {
         let nostr_group_id = create_test_nostr_group_id();
         let image_url = Some("http://blossom_server:4531/fake_img.png".to_owned());
         let image_key = Some(generate_encryption_key());
+        let image_nonce = Some(vec![16u8; 12]);
         let group = Group {
             mls_group_id: mls_group_id.clone(),
             nostr_group_id,
@@ -302,23 +305,14 @@ mod tests {
             state: GroupState::Active,
             image_url,
             image_key,
+            image_nonce,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let relay_url1 = RelayUrl::parse("wss://relay1.example.com").unwrap();
         let relay_url2 = RelayUrl::parse("wss://relay2.example.com").unwrap();
-        let group_relay1 = GroupRelay {
-            mls_group_id: mls_group_id.clone(),
-            relay_url: relay_url1,
-        };
-        let group_relay2 = GroupRelay {
-            mls_group_id: mls_group_id.clone(),
-            relay_url: relay_url2,
-        };
+        let relays = BTreeSet::from([relay_url1, relay_url2]);
         nostr_storage
-            .save_group_relay(group_relay1.clone())
-            .unwrap();
-        nostr_storage
-            .save_group_relay(group_relay2.clone())
+            .replace_group_relays(&mls_group_id, relays)
             .unwrap();
         let found_relays = nostr_storage.group_relays(&mls_group_id).unwrap();
         assert_eq!(found_relays.len(), 2);
@@ -333,11 +327,6 @@ mod tests {
                 panic!("Group relays not found in cache");
             }
         }
-        nostr_storage
-            .save_group_relay(group_relay1.clone())
-            .unwrap();
-        let found_relays_after_duplicate = nostr_storage.group_relays(&mls_group_id).unwrap();
-        assert_eq!(found_relays_after_duplicate.len(), 2);
     }
 
     #[test]
@@ -348,6 +337,7 @@ mod tests {
         let nostr_group_id = create_test_nostr_group_id();
         let image_url = Some("http://blossom_server:4531/fake_img.png".to_owned());
         let image_key = Some(generate_encryption_key());
+        let image_nonce = Some(vec![16u8; 12]);
         let group = Group {
             mls_group_id: mls_group_id.clone(),
             nostr_group_id,
@@ -360,6 +350,7 @@ mod tests {
             state: GroupState::Active,
             image_url,
             image_key,
+            image_nonce,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let group_exporter_secret_0 = GroupExporterSecret {
@@ -434,6 +425,7 @@ mod tests {
             group_description: "A test welcome group".to_string(),
             group_image_key: None,
             group_image_url: None,
+            group_image_nonce: None,
             group_admin_pubkeys: BTreeSet::from([pubkey]),
             group_relays: BTreeSet::from([RelayUrl::parse("wss://relay.example.com").unwrap()]),
             welcomer: pubkey,
@@ -494,6 +486,7 @@ mod tests {
         let nostr_group_id = create_test_nostr_group_id();
         let image_url = Some("http://blossom_server:4531/fake_img.png".to_owned());
         let image_key = Some(generate_encryption_key());
+        let image_nonce = Some(vec![16u8; 12]);
         let group = Group {
             mls_group_id: mls_group_id.clone(),
             nostr_group_id,
@@ -506,6 +499,7 @@ mod tests {
             state: GroupState::Active,
             image_url,
             image_key,
+            image_nonce,
         };
         nostr_storage.save_group(group.clone()).unwrap();
         let event_id = EventId::all_zeros();
@@ -590,6 +584,7 @@ mod tests {
         let nostr_group_id = create_test_nostr_group_id();
         let image_url = Some("http://blossom_server:4531/fake_img.png".to_owned());
         let image_key = Some(generate_encryption_key());
+        let image_nonce = Some(vec![16u8; 12]);
         let group = Group {
             mls_group_id: mls_group_id.clone(),
             nostr_group_id,
@@ -602,6 +597,7 @@ mod tests {
             state: GroupState::Active,
             image_url,
             image_key,
+            image_nonce,
         };
 
         // Save the group
@@ -623,6 +619,8 @@ mod tests {
         let nostr_group_id = create_test_nostr_group_id();
         let image_url = Some("http://blossom_server:4531/fake_img.png".to_owned());
         let image_key = Some(generate_encryption_key());
+        let image_nonce = Some(vec![16u8; 12]);
+
         let group = Group {
             mls_group_id: mls_group_id.clone(),
             nostr_group_id,
@@ -635,6 +633,7 @@ mod tests {
             state: GroupState::Active,
             image_url,
             image_key,
+            image_nonce,
         };
 
         // Save the group