+++ mls/nostr-mls-sqlite-storage/CHANGELOG.md
@@ -28,6 +28,7 @@
 ### Breaking changes
 
 - Remove group type from groups
+- Replaced `save_group_relay` with `replace_group_relays` trait method (https://github.com/rust-nostr/nostr/pull/1056)
 
 ### Changed